requirements.txt
requests
gevent
pyyaml
loguru
pytz
fire